requirements.txt
asgiref
autopep8
blinker
click
colorama
crispy-bootstrap5
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
djangorestframework
EasyProcess
entrypoint2
Flask
itsdangerous
Jinja2
Markdown
MarkupSafe
MouseInfo
mss
mysqlclient
Pillow
psycopg2
psycopg2-binary
PyAutoGUI
pycodestyle
PyGetWindow
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
PythonTurtle
pytweening
pytz
reportlab
sqlparse
typing_extensions
tzdata
Werkzeug
